int to long

+++ source/batch/src/main/java/org/maca/continuous/perftest/app/model/CountMetrics.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CountMetrics {
     public String name;
-    public Integer value;
+    public long value;
 }
